convert to typed reconciler - not working yet
