+++ pkg/services/featuremgmt/toggles_gen.json
@@ -2671,11 +2671,14 @@
     {
       "metadata": {
         "name": "playlistsReconciler",
-        "resourceVersion": "1730462910506",
-        "creationTimestamp": "2024-11-01T12:08:30Z"
+        "resourceVersion": "1734463170112",
+        "creationTimestamp": "2024-11-01T12:08:30Z",
+        "annotations": {
+          "grafana.app/updatedTimestamp": "2024-12-17 19:19:30.112629 +0000 UTC"
+        }
       },
       "spec": {
-        "description": "Enables experimental watcher for playlists",
+        "description": "Enables experimental reconciler for playlists",
         "stage": "experimental",
         "codeowner": "@grafana/grafana-app-platform-squad",
         "requiresRestart": true
